+++ include/utils/SparseLineGrid.h
@@ -1,20 +1,21 @@
-//Copyright (c) 2018 Ultimaker B.V.
-//CuraEngine is released under the terms of the AGPLv3 or higher.
+// Copyright (c) 2018 Ultimaker B.V.
+// CuraEngine is released under the terms of the AGPLv3 or higher.
 
 
 #ifndef UTILS_SPARSE_LINE_GRID_H
 #define UTILS_SPARSE_LINE_GRID_H
 
 #include <cassert>
+#include <functional>
 #include <unordered_map>
 #include <vector>
-#include <functional>
 
 #include "IntPoint.h"
-#include "SparseGrid.h"
 #include "SVG.h" // debug
+#include "SparseGrid.h"
 
-namespace cura {
+namespace cura
+{
 
 /*! \brief Sparse grid which can locate spatially nearby elements efficiently.
  *
@@ -36,17 +37,18 @@ public:
      * \param[in] elem_reserve Number of elements to research space for.
      * \param[in] max_load_factor Maximum average load factor before rehashing.
      */
-    SparseLineGrid(coord_t cell_size, size_t elem_reserve = 0U, float max_load_factor = 1.0f);
+    SparseLineGrid(coord_t cell_size, size_t elem_reserve = 0U, double max_load_factor = 1.0);
 
     /*! \brief Inserts elem into the sparse grid.
      *
      * \param[in] elem The element to be inserted.
      */
-    void insert(const Elem &elem);
+    void insert(const Elem& elem);
 
     void debugHTML(std::string filename);
 
     static void debugTest();
+
 protected:
     using GridPoint = typename SparseGrid<ElemT>::GridPoint;
     using grid_coord_t = typename SparseGrid<ElemT>::grid_coord_t;
@@ -56,31 +58,30 @@ protected:
 };
 
 
-
 #define SGI_TEMPLATE template<class ElemT, class Locator>
 #define SGI_THIS SparseLineGrid<ElemT, Locator>
 
 SGI_TEMPLATE
-SGI_THIS::SparseLineGrid(coord_t cell_size, size_t elem_reserve, float max_load_factor)
- : SparseGrid<ElemT>(cell_size, elem_reserve, max_load_factor)
+SGI_THIS::SparseLineGrid(coord_t cell_size, size_t elem_reserve, double max_load_factor)
+    : SparseGrid<ElemT>(cell_size, elem_reserve, max_load_factor)
 {
 }
 
 SGI_TEMPLATE
-void SGI_THIS::insert(const Elem &elem)
+void SGI_THIS::insert(const Elem& elem)
 {
     const std::pair<Point, Point> line = m_locator(elem);
     using GridMap = std::unordered_multimap<GridPoint, Elem>;
     // below is a workaround for the fact that lambda functions cannot access private or protected members
     // first we define a lambda which works on any GridMap and then we bind it to the actual protected GridMap of the parent class
-    std::function<bool (GridMap*, const GridPoint)> process_cell_func_ = [&elem, this](GridMap* m_grid, const GridPoint grid_loc)
-        {
-            m_grid->emplace(grid_loc, elem);
-            return true;
-        };
-    using namespace std::placeholders;  // for _1, _2, _3...
+    std::function<bool(GridMap*, const GridPoint)> process_cell_func_ = [&elem, this](GridMap* m_grid, const GridPoint grid_loc)
+    {
+        m_grid->emplace(grid_loc, elem);
+        return true;
+    };
+    using namespace std::placeholders; // for _1, _2, _3...
     GridMap* m_grid = &(this->m_grid);
-    std::function<bool (const GridPoint)> process_cell_func(std::bind(process_cell_func_, m_grid, _1));
+    std::function<bool(const GridPoint)> process_cell_func(std::bind(process_cell_func_, m_grid, _1));
 
     SparseGrid<ElemT>::processLineCells(line, process_cell_func);
 }
@@ -89,13 +90,13 @@ SGI_TEMPLATE
 void SGI_THIS::debugHTML(std::string filename)
 {
     AABB aabb;
-    for (std::pair<GridPoint, ElemT> cell:  SparseGrid<ElemT>::m_grid)
+    for (std::pair<GridPoint, ElemT> cell : SparseGrid<ElemT>::m_grid)
     {
         aabb.include(SparseGrid<ElemT>::toLowerCorner(cell.first));
         aabb.include(SparseGrid<ElemT>::toLowerCorner(cell.first + GridPoint(SparseGrid<ElemT>::nonzero_sign(cell.first.X), SparseGrid<ElemT>::nonzero_sign(cell.first.Y))));
     }
     SVG svg(filename.c_str(), aabb);
-    for (std::pair<GridPoint, ElemT> cell:  SparseGrid<ElemT>::m_grid)
+    for (std::pair<GridPoint, ElemT> cell : SparseGrid<ElemT>::m_grid)
     {
         // doesn't draw cells at x = 0 or y = 0 correctly (should be double size)
         Point lb = SparseGrid<ElemT>::toLowerCorner(cell.first);
@@ -112,7 +113,7 @@ void SGI_THIS::debugHTML(std::string filename)
             lb.Y = -SparseGrid<ElemT>::cell_size_;
             rb.Y = -SparseGrid<ElemT>::cell_size_;
         }
-//         svg.writePoint(lb, true, 1);
+        //         svg.writePoint(lb, true, 1);
         svg.writeLine(lb, lt, SVG::Color::GRAY);
         svg.writeLine(lt, rt, SVG::Color::GRAY);
         svg.writeLine(rt, rb, SVG::Color::GRAY);
